Sif Rework P3: Add a Divine Knowledge ability
Sif currently has one active that uses piety, Channel Energy, and this
is an ability that becomes more about convenience than power as the game
progresses since players get access to a large MP pool. This commit adds
a powerful new active to help Sif be relevant at the middle and late
portions of the game and to help cement her role as the more active
magical god counterpart to the more passive Vehumet.

Tentatively named Divine Knowledge, this ability allows casting any
spell in the player's library with 100% success and with a spell power
bonus of 1.5 x Invocations skill in addition to any spell-power coming
from Int, Spellcasting, and any training in the spell schools of the
spell. Thus players can cast a spell that normally has a 100% fail rate
due to no or little training in the relevant skills. Divine Knowledge is
granted at 5* and has a high piety cost of 12, meaning several
uses of the ability is likely to bring a player under 5*. The ability
will certainly be useful to a wide variety of spell using characters,
granting limited access to high level spells that normally remain
totally inaccessible behind high skill requirements.

This gift timeout from gifts starting at 5* means it's very easy to drop
from 5* to 4*, losing access to the ability. It might be best to either
change the numbers behind the timeout and gift frequency, which
potentially has the side benefit of allowing us to move book gifting
earlier, or to move Divine Knowledge to 4* (but still with the same high
peity cost). We'd like to try granting the gifts a bit earlier, but it
doesn't work well in practice to have have a long gift timeout eating
piety at a lower piety level and a high cost active granted at a higher
piety levels. Quite long stretches of the game can happen without the
player getting past the gift timeout.

Play-testing with the Sif rework commits has shown that having both
Divine Knowledge and gifts at 5* does give some access to both, so this
setup can do for now.