Make Divine Knowledge fail upon spell selection
Previous to this commit, trying to activate the ability would check the
fail chance roll, as is done for most god abilities. This blocked access
to the spell listing, which is inconvenient since players can't see what
they might be able to cast in terms of available spells and their
potential spellpower. This commit has the fail check happen when the
spell is selected.

It might be best to have the fail checked after the actual spell
targeting is done, so players can use any spell targeting to see what
they might be able to hit. This has a drawback where the player has to
do the work of selecting a target repeatedly if they have one or more
failures. If we decide that having the fail check happen after target
selection is done, some of the spellcasting functions can be
appropriately modified to have a fail check variable, notably
cast_a_spell() and your_spells().