Fedhas rework part 6: simplify ballistomycetes
Ballistomycetes are complicated monsters that grow based on interactions
with the spore monsters. Spores leave mold as they move, which can give
rise to the production of new ballistomycetes. This set of behaviors
fell into the realm of "complexity for the sake of complexity". One had
to wait a while for this "reproduction" to play out, only to get a few
more ballistomycetes spores. Now that we're putting stronger limits on
ally usage through use of summons, it's a good time to simplify the way
ballistomycetes and their spores work.

This commit removes the reproduction aspect of ballistomycetes, removes
mold itself, and simplifies ballistomycetes down to one type of monster.
Additionally, ballistos now have a Sporulate natural ability they
occassionally use to create a single spore as a short-lived summon. The
spore will seek after the ballisto's target and explode upon hit, still
producing the damaging explosion that applies confusion to living
targets.

A subsequent commit will make a Fedhas active ability to let the player
to create a ballistomycete summon, and this will replace Fedhas'
Reproduction.