Fedhas rework part X: vines passive (disabled)
This is a (currently disabled) implementation of a passive halo that's
somewhat similar to Ledas' Liquefaction. It increases movement energy
costs by 3 for all monsters in the halo. It's themed as vines that
hinder the movement of all substantial, non-flying monsters. If enabled,
it would start with radius 1 at 1* piety, increasing by 2 (for
non-barachi) until reaching LOS radius at 4* piety.

This type of passive likely won't make it into the final rework since
the idea has a lot of problems. It increases kiting and directly
counteracts the movement limitations of slow species. Directly
counteracting signification species limitations is something we do with
other gods, but that's not to say we should increase this. Still, the
basic halo implementation might be useful if we do find a good passive,
and if not, this commit can easilly be rebased out.

The halo overlay in tiles is a transparent green overlay, and in console
it colors the floor lightgreen. The lightgreen floor color is used only
in Snake Pits, so it might be good to change the floor color in that
branch if this halo makes it in.