Fedhas rework part 9: Overgrow ability
A 4* ability that transforms any (non-permarock) walls, trees, doors, or
diggable solid features in a 3x3 smite-targeted area into plant allies
at a cost of 12 piety. The created monster is a summon, but the wall is
permanently destroyed.

This commit implements a targeter UI that only shows cells as affected
if they contain a wall that's a valid target for Overgrow. To be
affected, a wall must be in LOS and not blocked by a feature that blocks
translocations (e.g. clear walls). Open doors won't be affect if they
contain a monster, and if it turns out there was an invisible monster in
the open door, the ability will fail but still use piety.

Currently the set of possible plant allies is weighted towards oklob
saplings, wandering mushrooms, and burning bushes, with lower chances
for ballistomycetes and oklob plants. The plant HD is increased by 1
with each level of Invocations, which for oklob saplings, burning
bushes, and oklob plants will give more damage per use of their
abilities, and oklob plants will get an increase in their "spit" rate.
All monsters will get increased HP and summon duration with Invocations.