Add text entry widget
This commit adds a rudimentary textbox widget. It is single-line only,
and does not support text selection or, indeed, most of the niceties of
regular OS/toolkit-provided textboxes. That can come later.

This widget has an inner LineReader class, which is a gutted version of
the line_reader class, permitting independent modification. line_reader,
meanwhile, will hopefully eventually be removed as more and more UI
layouts are widgified.