Fix focus traps breaking webtiles chat
Focus traps on the top-level popup prevent the chat box from being
focused. To fix this, when showing the chat, we enter a new focus trap,
and leave it when returning to the game. This is complicated by the fact
that the chat JS is effectively non-modifiable (not in game_data), so we
attach to the focusin event. You can tab around within the chat popup;
the focus-trap class prevents nested traps when doing so.