Encode nested colours in formatted_string
This attempts to make formatted_string a closer fit to the textual
colour string format; nested colours are now encoded, and parsing a
colour string and then converting to a string again should in many cases
now be a no-op.

parse_string_to_multiple() is also changed so that the colour nesting is
preserved across newlines, i.e. the colour stack is unwound and then
rewound at each newline.