An icy wall dungeon feature
The feature will be used in a forthcoming Frozen Ramparts spell, where
it will damage and debuff monsters that walk adjacent to the wall in a
way similar to the slimy walls used in the Slime Pits. This commit adds
the feature data, description, and graphical handling. A new level state
flag is set whenever an icy wall exists that allows the floor colors and
tiles to change and to draw the overlays. Some future refactoring might
obviate the need for this flag as these walls are likely to only be set
on a temporary basis by the spell that creates them.

The tiles are color rotated and modified version of the slimy walls,
slimy wall overlays, and acid floors so that they look frozen and icy.

The wall effect will be implemented as part of the spell in a subsequent
commit.