Frozen Ramparts, an L3 Ice/Hexes spell
This spell is available to Ice Elementalists in their starting Book of
Frost. Frozen Ramparts temporarily changes all diggable rock walls in
radius 3 of the player into icy walls that damage and debuff monsters
that walk adjacent to them. The damage is similar to freeze, although
this spell has a higher spellpower cap of 50. The debuff is the Frozen
status that slows monster movement (but not other actions).

The player and allies are not affected by these walls and can move
safely adjacent to them. Player LOS of the monster is required for the
walls to have an effect, and they revert to normal rock when LOS of the
wall is broken or when the Ramparts status expires. Spellpower increases
the duration of the walls as well as the damage. The damage uses
BEAM_ICE, hence is only 40% resistible, yet it does apply AC, unlike
Freeze. This is themed as a release of a blast of icicles when a victim
walks by, encasing them in ice.

This commit is a work in progress, and details of the spell including
damage numbers, the type of debuff, and other aspects of the walls could
change before or after it's merged to trunk.